Fix ccache permissions
For some reason, the ownership of build_user_home/.ccache gets set to
root:root on some platforms, but not on others.  I don't see a directory
resource block for it anywhere.  Assuming it must be created during the
Chef run.

This should walk the build_user_home/.ccache and correct the ownership,
so that `omnibus build $project` can use build_user_home/.ccache.